Move to own package.Removed fallback as Node 16 is now lowest node version.
